Avoid accessing a nil pointer while running update_swissindex_pharma
